remove assertions file
